AI changes to Red, Blue and Twins + Red's team has been revamped
The Twin's AI was stupidly basic (which is understandable, because they're kids, but still), so I added at least the AI_BASIC constant.

Red and Blue are a little more annoying thanks to AI_OPPORTUNIST. If you're low in HP, heal yourself because God won't give you a second chance.

Finally, Red's team items and moveset have been slightly changed to be more challenging; after all, he's the final boss of the game.