Time-Weather based heal is now only WEATHER-based Heal
Moves that used to heal based on both time of the day and weather (Morning Sun, Synthesis and Moonlight) now only take the WEATHER into account. Also, the amount of healing has been adjusted to match those of recent generations (1/4, 1/2 and 2/3 Max HP).

With this change, the code has been slightly cleaned, too.