fix: Get tile permissions properly from final tile
Explanation: Tile permissions (collisions) are retrieved for the final tile. This is a problem in stairs because when it gets the permissions at the start of the 2nd "horizontal step" out true Y-position hasn't been updated; correcting the position doesn't trigger another tile permission check.

The solution is to update our final Y-axis map position before triggering the 2nd horizontal step, which will give us the correct final tile permissions.